requirements.txt
python-telegram-bot==20.3
web3==6.10.0
